Move internal files into correct packages
